refactor(configcheck): move the abnormal-failure retry to the shared package
Both aggregators carried a byte-identical copy of the branch that decides a
failed configcheck pod was killed around the check rather than by an invalid
config, deletes it, and reports the retry - plus a copy of the
stillRunningContainer helper behind its log line. A later fix to those
semantics would have landed in one copy only.

pkg/resources/configcheck already owns the concerns both configchecks share,
so the branch moves there and each caller keeps only the result mapping it
owns. No behaviour change: same condition, same log, same delete, same
message.
